+++ R/RcppExports.R
@@ -21,6 +21,10 @@ crfsuite_model_dump <- function(file_model, file_txt) {
     invisible(.Call('_crfsuite_crfsuite_model_dump', PACKAGE = 'crfsuite', file_model, file_txt))
 }
 
+crfsuite_model_coefficients <- function(file_model) {
+    .Call('_crfsuite_crfsuite_model_coefficients', PACKAGE = 'crfsuite', file_model)
+}
+
 crfsuite_predict <- function(file_model, doc_id, x, trace = 0L) {
     .Call('_crfsuite_crfsuite_predict', PACKAGE = 'crfsuite', file_model, doc_id, x, trace)
 }